v4l2/stateless: Frame id normalization
The v4l2_buffer->timestamp field is the source of truth. |timestamp|
is a struct timeval. The member |tv_usec| is used to hold the frame id
and is of type uint64_t.

Cq-Include-Trybots: luci.chromium.try:linux-v4l2-codec-rel
Bug: b:278935312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147350
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240360}

+++ media/gpu/v4l2/stateless/stateless_decode_surface.cc
@@ -10,7 +10,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace media {
 
-StatelessDecodeSurface::StatelessDecodeSurface(uint32_t frame_id,
+StatelessDecodeSurface::StatelessDecodeSurface(uint64_t frame_id,
                                                base::OnceClosure enqueue_cb)
     : frame_id_(frame_id), enqueue_cb_(std::move(enqueue_cb)) {
   DVLOGF(3) << "Creating surface with id  : " << frame_id_;
